Algorithmic-Trading-System/requirements.txt
numpy==1.24.3
pandas==2.0.3
yfinance==0.2.18
ta-lib==0.4.26
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1
jupyter==1.0.0
pytest==7.4.0
black==23.7.0
flake8==6.0.0
alpha-vantage==2.3.1
requests==2.31.0
python-dotenv==1.0.0
backtrader==1.9.78.123
zipline-reloaded==2.2.0
pyfolio==0.9.2
empyrical==0.5.5
quantstats==0.0.59
arch==5.6.0
scipy==1.11.1
tensorflow==2.13.0
keras==2.13.1
torch==2.0.1
torchvision==0.15.2
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2
optuna==3.3.0
shap==0.42.1
lime==0.2.0.1
hyperopt==0.2.7
mlflow==2.5.0
wandb==0.15.8
joblib==1.3.2
cvxpy==1.4.1
cvxopt==1.3.2
PyPortfolioOpt==1.5.5
riskfolio-lib==4.0.1
zipline-reloaded==2.2.0
fredapi==0.5.1
yfinance==0.2.18
quandl==3.7.0
beautifulsoup4==4.12.2
lxml==4.9.3
websocket-client==1.6.3
dash-bootstrap-components==1.5.0
dash-daq==0.5.0
flask==2.3.3
flask-socketio==5.3.6
gunicorn==21.2.0
redis==4.6.0
celery==5.3.1
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
aiofiles==23.2.1
asyncio==3.4.3
websockets==11.0.3
flask-cors==4.0.0